+++ src/utils/mod.rs
@@ -3,3 +3,5 @@ pub mod sse;
 
 #[cfg(feature = "google")]
 pub mod json_schema;
+
+pub mod params;